move source code display to a reuseable component
